+++ Sources/API/Model/Endpoints.swift
@@ -85,10 +85,14 @@ enum Endpoints {
             typealias EndpointType = WPNEndpointBasic<WPNRequest<VerifyOTPRequest>, WPNResponse<VerifyOTPResponse>>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/otp/verify", e2ee: .activationScope) }
         }
+        enum ActivationFinish {
+            typealias EndpointType = WPNEndpointSignedWithToken<WPNRequest<ActivationFinishRequest>, WPNResponse<ActivationFinishResponse>>
+            static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/activation", tokenName: "possession_universal", e2ee: .activationScope) }
+        }
     }
     enum Configuration {
         enum GetConfiguration {
-            typealias EndpointType = WPNEndpointBasic<WPNRequest<ConfigurationRequest>, WPNResponse<ConfigurationResponse>>
+            typealias EndpointType = WPNEndpointBasic<WPNRequest<WDOConfigurationRequest>, WPNResponse<WDOConfigurationResponse>>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/configuration", e2ee: .applicationScope) }
         }
     }